backend/requirements.txt
# Web Dependencies
eventlet==0.33.0
python-socketio==5.5.1
websocket-client==1.2.3
flask==2.0.2
flask-socketio==5.1.1
flask-cors==3.0.8

# Controller Support Dependencies
inputs==0.5